ltp-aiodio: aiodio_sparse fix race condition
On slower machines one of the forked childs may stay waiting to be
scheduled (due to heavy load the main testing process generates) before
it has chance to change it's signal handler so when parent starts
killing the childrens it has still the ltp signal handler which results
in test failure. Fix this by temporarily changing the signal handler
before the main process starts forking.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>