getdents: rewrite syscall handling completely
The inline asm has many problems, but rather than attempt to fix
them, just use syscall() for everyone.  This allows us to drop the
i386-specific checks and have the tests run on all arches.

Further, add a layer between the kernel and the dirent struct that
the tests uses.  The kernel packs the results, so we need to expand
the raw buffer returned by the kernel into the userland structs we
pass around.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>