pthread_join/6-3: split the do_it switch
If the "creating thread" thread is dead before the "sem_wait"
thread, then the case will hang up on "sem_wait" forever .
So, we split the switch to do_it1 and do_it2, and make sure
that we can stop the "sem_wait" thread before the "creating thread"
thread.

And remove the sleep(10) since we don't need it after splitting
the do_it switch now.

Signed-off-by: Wanlong Gao <gaowanlong@cn.fujitsu.com>