Renamed /lib/str_to_bytes.c and /include str_to bytes.h
The str_to_bytes files have been renamed to bytes_by_prefix in order
to make it easier for anyone to understand the purpose of the file,
the functions of the file have also been ranmed accordingly.

Furthermore any call to the file or its functions by other functions,
manpages etc. have been modified to comply with the new names.

Signed-off-by: Marios Makris <marios.makris@gmail.com>
Reviewed-by: Wanlong Gao <gaowanlong@cn.fujitsu.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>