timer_settime/5-3: fix test hung
This case hangs on sugarbay(intel x86_64) platform.

The root cause is that the case calls timer_settime with time which
already took place and then calls sleep. When the timer expires and
sends signal the sleep will be break.

But on sugarbay the timer always expires before the case enters sleep,
then it always sleeps NUMTESTS * LONGSLEEPTIME seconds. So it exceeds
the time limit set by auto-run test shell script, and reports that it
hangs.

Decrease LONGSLEEPTIME value to avoid this issue.

Signed-off-by: Kang Kai <kai.kang@windriver.com>
Reviewed-by: Jan Stancek <jstancek@redhat.com>
Signed-off-by: Wanlong Gao <gaowanlong@cn.fujitsu.com>