open_posix_testsuite/.../mmap/5-1.c: Fixes.
This removes the ifdef PROT_XXX because if
these are not defined the test would fail
to compile due to undeclared symbols and
these messages wouldn't be printed anyway.

The second change improves coverage for the
test simply by trying all combinations of
prots and flags that are permitted by POSIX.
Each such combination must either succed
or fail with ENOTSUP any other result is
treated as a failure.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>