open_posix_testsuite/.../mmap/12-1.c: Rewrite.
The original test was wrong and passed only
because of (possibly intentional) typo.

Now the tests removes all references to the
file but the mapping and contines to
read/write/sync it. Then it's just unmapped,
unfortunately we have no way how to tell that
unlinked file was really removed.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>