thp01: make "true" applet foundable for busybox
Many embedded systems use busybox. busybox only has one binary, so it depends
on argv[0] as the applet name to execute right command. in thp01 case, it is
"true" command.
this fixes the issue thp01 case always reports "applet not found" for embedded
systems and make the case pass on those systems using busybox.
At the same time, this fix doesn't destory the validation of the test case.

Signed-off-by: Barry Song <Barry.Song@csr.com>
Signed-off-by: Wanlong Gao <gaowanlong@cn.fujitsu.com>