Add easy to use API for parent-child synchronization.
There are various places in LTP where execution of
parent and child must be synchronized. Parent must
run before the child or child must run before the
parent. These situations are now handled by calls
to sleep so that kernel scheduller has chance to
run the process we need.

This commit adds easy to use API to handle this
problem correctly. It makes use of fifo as IPC
mechanism which was used over pipe because pipe
is not usable with LTP on uClinux.

See lib/tests/tst_checkpoint_*.c for example usage.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>