Add TST_RESOURCE_COPY() to prepare test files.
Currently there is no estabilished way how to
access additional files needed for testing.
This commit adds TST_RESOURCE_COPY() interface
that allows you to copy files to the test working
directory.

The function looks either for files in current
directory (actually the directory test was
started in) or LTPROOT or path "/opt/ltp" as
last resort. This allows you to run the test
both from the three it was compiled or from
installed LTP.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>