syscalls/getdents: Add parameter to test getdents64 syscall
Add new "-l" parameter to test the getdents64 syscall along with
the getdents one. Use ltp_syscall() to call the system call to get
proper return values when an architecture does not support any of the
getdents{,64} syscalls. Modify runtest/syscalls to call getdentsXX tests
with and without the "-l" parameter

Signed-off-by: Markos Chandras <markos.chandras@imgtec.com>