open_posix_testsuite/.../aio_suspend/9-1.c
Fix the testcase to adjust aio request sizes as runtime.

This testcase fails on machines that are fast enough to complete
the hardcoded request size before we get to the aio_suspend(). This
patch changes the testcase to do several tries with increasing buffer
sizes in order to cover both small embedded systems and large
mainframes.

There were several race conditions in the original code that had to be
fixed in order to be able to run the test repeatedly.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>