open_posix_testsuite/.../lio_listio/11-1.c: Remove.
The testcase is invalid as POSIX doesn't define what 'successfully
queued' means. On Linux (with glibc) 'successfully queued' means that
the operation was added to the list of aio requests to be dealt with
later which is in accordance with POSIX which request the lio_listio()
with mode LIO_NOWAIT to return immediately when the requests are queued.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>