sendmsg02: fix compilation on kernel < 2.6.27
Using SOCK_CLOEXEC and SOCK_NONBLOCK breaks compilation
on older distros/kernels. These were originally taken from
systemd sources, but as it turns out they are not needed to
trigger this issue.

I confirmed, that issue can still be reproduced on 3.8 kernel
and that testcase can be compiled at least back to RHEL5.3 (2.6.18).

Reported-by: DAN LI <li.dan@cn.fujitsu.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>
Acked-by: Wanlong Gao <gaowanlong@cn.fujitsu.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>