Fix retval parameter to pthread_join()
The retval parameter to pthread_join() is a pointer to a pointer. There
are several instances where a pointer to int is being passed instead.
This can result in a bus error on sparc64 when the pointer is not
aligned on a 64-bit boundary. Of course, it's a potential problem on
any 64-bit platform.

Also removed an unnecessary malloc & free from mmstress.c

Signed-off-by: Dave Kleikamp <dave.kleikamp@oracle.com>