tst_run_cmd: Make the tst_run_cmd survive SIGCHLD
The tst_sig() installs poisoned signal handlers for all signals the test
is not expected to get, this causes abort in tst_run_cmd right after the
child exits and parent starts running.

This commit makes the tst_run_cmd temporarily disable the handler for
SIGCHLD (the handler is set to SIG_DFL at the start of the tst_run_cmd
function and restored at the end of it) so that we don't have to disable
the poisoned handler in each test using tst_run_cmd.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>