lib/mem: handle missing online file
gather_node_cpus() fails if the /sys/devices/system/node/nodeX/cpuX/online
file is missing for any cpu other than cpu0. The absence of the online file
should not be treated as a failure, but as an indication that the cpu is
not hot-pluggable and cannot be taken offline.

Signed-off-by: Dave Kleikamp <dave.kleikamp@oracle.com>
Reviewed-by: Wanlong Gao <gaowanlong@cn.fujitsu.com>
Reviewed-by: Jan Stancek <jstancek@redhat.com>