controllers/cpuset: required number of nodes
Initially these testcases required a host with 4 CPUs
and 3 memory nodes minumum. This prevented to run the
testcases on widely used 2-memory-nodes configuration.

And in fact the testcases only need 2 memory nodes, and
a few needs 4 CPUs.

Modified the checks accordingly. Now "ideal" minumum
configuration of a host is 4 CPUs and 2 memory nodes.

Signed-off-by: Stanislav Kholmanskikh <stanislav.kholmanskikh@oracle.com>