controllers/cpuset: Fix cleanup.
On some systems (found on SLES10) cpuset mounts aren't listed in the output
from mount (they are not in /etc/mtab at all), so we switch to the /proc/mounts
instead.

Also fix a race in the loop that kills processes that are left in the currently
mounted cpuset. The problem is that if you kill a LTP test, the cleanup is
still executed and the test takes some time to finish which results in EBUSY in
the umount that is executed right after the loop. The fix is to kill the
processes with -9 (SIGKILL) which makes them exit immediately.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>