getcontext01: check getcontext function in libc
getcontext function is implemented depends on libc version and arch.
Some libc version for specific arch does not implement getcontext yet.
In that case, getcontext just sets errno to ENOSYS and returns -1.
For example, getcontext is not implemented before libc-2.17 in ARM.
Current test goes to TFAIL when lower version of libc is used in ARM.

This patch sets the test as TCONF if libc does not implement getcontext.
If ENOSYS is not set and -1 is returned, it goes to TFAIL as it was.

Signed-off-by: Honggyu Kim <hong.gyu.kim@lge.com>