sendfile08: fix buf comparison
The sendfile08 testcase uses the sendfile() syscall
to write the strings "hello" and "world" to a file descriptor
and then reads them back in through another file descriptor.

The length of the strings written are determined by strlen(),
so that in total 10 bytes are written.

When reading the strings back in, again 10 bytes are read, but
the problem is, that those are read back into a *local* string
buffer array (buf[]) which hasn't been initialized.

Afterwards,
  strcmp(buf, "helloworld")
is called to compare input and output.
This strcmp() is wrong, because basically "helloworld\0"
(with the trailing string zero terminator) is compared to
"helloworld" (likely without a trailing zero in the buf[] array).

So, if buf[] hasn't been initialized, it's not guaranteed
that the byte in buf[10] equals to '\0'.

This problem was found on hppa64-linux-gnu architecture, most
likely, because it's a stack-grows-up architecture.

Fix this by using strncmp().

Signed-off-by: Helge Deller <deller@gmx.de>
Signed-off-by: Jan Stancek <jstancek@redhat.com>