lib: Add tst_process_state_wait2()
The tst_process_state_wait() uses the tst_brkm() interface and is not
suitable for use in child processes.

This adds similar tst_process_state_wait2() function which does the same
but does not use tst_brkm() to exit on failure. Instead of that it
returns non-zero exit value and prints the cause of the failure into the
stderr.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>