openposix: Fix shm_open, shm_unlink ENAMETOOLONG
The ENAMETOOLONG was changed to may fail (was previously shall fail) in
latest POSIX (since 2008). So ENAMETOOLONG is no longer required error
and since glibc commit (included since glibc 2.19) these testcases
started to fail.

commit 5d30d853295a5fe04cad22fdf649c5e0da6ded8c
Author: Ondrej Bilka <neleai@seznam.cz>
Date:   Thu Oct 31 13:58:01 2013 +0100

    Restrict shm_open and shm_unlink to SHMDIR. Fixes bugs 14752 and 15763.

This fix changes the testcases to return UNTESTED if the POSIX version
is decleared newer than 2008 and the call has failed with different
error than ENAMETOOLONG.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>