fork/fork14.c: cleanup
If mmap() in fork_test() fails too many times, make this test return
TCONF, because in this condition, we are almost impossible to get an
vm_area_struct sized 16TB, that is though this bug may exist, the
overflow would not occur.

According to this url:
http://lkml.iu.edu//hypermail/linux/kernel/1204.3/00416.html, which
also attached a reproducer program, make some code logic adjustments.
If at the beginning, as the program continues to mmap, the fork will
fail. And it continues to mmap, if at some point, fork() returns
success, then we can say the overflow occurs.

Signed-off-by: Xiaoguang Wang <wangxg.fnst@cn.fujitsu.com>