lib: Remove tst_uid_gid and cases that use it.
This commit removes the code that tries to find unused uid/gid because:

The kernel does not care if UID/GID is used in userspace or not; the
result is always either success or EPERM (unless you played with user
namespaces, in which case you can get EINVAL if there is no mapping
present, which is not the case here).

This makes the tests that use GET_UNUSED_* setregid02 and setresuid03
redundant and not worth of keeping.

Note that the original problem was that the value hardcoded for these
tests clashed with nobody UID/GID on SUSE distributions which made the
test case succeed unexpectedly.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>