utils/Makefile: Move ffsb configure to configure.ac
The configure call was done from Makefile without appropriate parameters,
which breaks cross-compiling. Include ffsb into configure.ac so that it is
included in the main configure call.

This fix is based on a patch by Fathi Boudra <fathi.boudra@linaro.org>
for an earlier version of LTP.

This also fixed subsequent makes for out-of-tree build that previously
failed when attempting to run ffsb configure script for the second time.

CC: Fathi Boudra <fathi.boudra@linaro.org>
Signed-off-by: Mats Liljegren <mats.liljegren@enea.com>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>