syscalls/fallocate: Fix regression.
Since we use the libc fallocate syscall wrapper (if avaiable) these
testcases started to fail on ix86. The reason is that we need to define
_GNU_SOURCE in order to get fallocate function prototype and that the
test uses loff_t to pass fallocate parameters (and fallocate() takes
off_t as parameters).

Now on ix86 (and possibly other 32bit platforms) without the right
function prototype the parameters are passed as they are (without
conversion) and loff_t is 64 bit in size while the function expects
off_t which is 32 bit. This results into wrong parameters passed to
the syscall (lenght was zero in my case) which yielded EINVAL and
the testcases failed.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>