pitest-2, pitest-3: increased the timeout
On some platforms do_work_dummy() (if compiled without any optimizations,
and it's by default) executes longer than the specified timeout of 100 ms.

Therefore there is a very small chance that TL may lock the mutex during
sleep() in the main thread, and most of the time the tests will fail with:
../../../functional/threads/pi_test/pitest-2.c: 222: FAIL: Thread TB1: lock returned 0 Success, slept 0.000008

To overcome this I increased the timeout value for TBs. It doesn't break the logic.

Signed-off-by: Stanislav Kholmanskikh <stanislav.kholmanskikh@oracle.com>
Acked-by: Jan Stancek <jstancek@redhat.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>