sigaction 9-1, 10-1, 11-1: use SIGKILL instead of SIGHUP
There are test environments, where the SIGHUP signal may be either
added to the signal mask or ignored. One example of such environments
is Autotest, which sets the SIGHUP signal to SIG_IGN for its jobs.

Therefore the above test cases hang in that environment because their
parent processes send SIGHUP to terminate the child (and the child ignores it).

I think it will not be a problem to use SIGKILL here instead of SIGHUP, because
SIGKILL can not be caught, blocked, or ignored.
And it does the job - stopping the child - very well :)

Signed-off-by: Stanislav Kholmanskikh <stanislav.kholmanskikh@oracle.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>