kernel/mem/lib/mem.c: make tests continue to run even though KSM is not enabled
When kernel KSM configuration is not enabled, we make testcases continue to run
instead of returning TCONF, e.g. oom01. And if KSM configuration is not enabled, in
the original code, "tst_brkm(TCONF, NULL, ...)" will be called, in this case, we
will forget to execute the restore operations defined in cleanup, which will forget
to restore the /proc/sys/vm/overcommit_memory to 0, and this will impact other test
cases in LTP, for example, fork14 in RHEL5 will succeed unexpectedly.

Signed-off-by: Xiaoguang Wang <wangxg.fnst@cn.fujitsu.com>
Acked-by: Stanislav Kholmanskikh <stanislav.kholmanskikh@oracle.com>
Signed-off-by: Wanlong Gao <gaowanlong@cn.fujitsu.com>