lib, openposix: Fix mount flags detection code.
The problem is that there are two records for root filesystem in the
/proc/mounts. One real and one dummy that looks like:

rootfs / rootfs rw 0 0

Now the problem is that the code matched the first one that has
incomplete flags and returned wrong results.

Previous attempt to fix this was to ignore all pseudo filesystem that
does have a path to a device in mnt_fsname.

That however excludes tmpfs that is commonly used for /tmp filesystem
nor it works for nfs mounts that starts with IP address.

This patch changes it to skip only the problematic entry with "rootfs"
in the mnt_fsname.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>