Remove STD_FUNCTIONAL_TEST
Only small part of the testcases was using STD_FUNCTIONAL_TEST and these
that were actually using it mostly misused the interface or used it only
to disable the tst_rest(TPASS, "sycall passed") message.

The default value was STD_FUNCTION_TEST enabled, so this only removes
code that is not executed in LTP testrun and there should be no
functional changes.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>