lib: Remove -c option and STD_COPIES
Remove -c option that was never working correctly.

The -c option, if set, forks the test in test setup() creating several
instances of the test to be running at the same time. The problems with
the implementation were:

* The failure of the forked tests WAS NOT popagaged to the first
  instance

* The test driver (ltp-pan) tracks only the first instance and kills the
  rest of the tests instances once the first one finishes

* Most of the tests does not expect to be running concurently
  (there are races in using the same test directory, loop devices, etc.)

While the first two would be easy to fix the last one would require
reviewing and testing all testcases which is too much work.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>