syscalls/rt_sigprocmask01: Fix race.
Since the sig_count was declared static the compiler is more eager in
optimization (it does not expect the value to change because it cannot
see the signal handler side effect) which causes the test to fail.

Declare the sig_count as volatile sig_atomic_t which fixes the failure
on x86_64.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>