sendfile09: regression test for 64bit sendfile capped at 2G
There is a bug on 64bit sendfile() with large file, please see
https://lists.ozlabs.org/pipermail/linuxppc-dev/2014-June/118128.html
for details.

This sendfile09.c was copied from sendfile02.c and using plain
file as the out_file. The file size is 4G (4294967296 bytes).

The count cannot be larger than 2G, but the offset can. So the second
case tests with 3G offset and 1G count. Many thanks to Jan Stancek's
help.

Signed-off-by: Han Pingtian <hanpt@linux.vnet.ibm.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>