remove meaningless TRETR flag
Flag TRETR was introduced to indicate test case was no longer valid and had
been "retired". But if you grep "TRETR" in LTP, you will not find any test
case using this flag. And if test case is no longer valid, we can just remove
it form LTP, so remove this flag.

Signed-off-by: Xiaoguang Wang <wangxg.fnst@cn.fujitsu.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>