profil01: rewrite testcase
The original testcase failed on ppc64le, because instead of using
_start address it reads value stored there, which caused that
calculated size of watched range (lotext, etext) was negative.

profil() syscall is deprecated for at least decade, so this really
excersises only glibc function. Since this test is in LTP for long
time and is quite simple, I chose to rewrite it, so we keep coverage.

Main changes are:
 1. no arch ifdefs, test assumes that 16k buffer will cover all code
    in test function
 2. only single sanity test to check that profil records some data

Signed-off-by: Jan Stancek <jstancek@redhat.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>