syscalls/keyctl01: Fix conditional compilation.
It looks like the configure check for keyctl was not working at all, at
least my logs shows that the test returned TCONF for previous release
too.

This patch removes the LTP_CHECK_KEYCTL_SYSCALL from configure.ac

(the m4 file was removed by accident by commit:

 commit 7c631b2d1145b999b92217b1f7b54431c19b9c53
 Author: Cyril Hrubis <chrubis@suse.cz>
 Date:   Tue Aug 19 09:50:06 2014 +0200

    configure, add_key{01,02}: Add check for linux/keyctl.h

Sorry everyone)

The test now makes use of the LTP_LINUX_KEYCTL_H which was introduced by the
prevously mentioned commit which enables the testcase correctly.

It also removes the keyutils.h header and uses int32_t for the key type, which
simplifies the the code (no need for just another configure check) and enables
the test on more targets where keyutils.h is not present.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>